refactor: 音声録音処理を他でも使えるように
