+++ src/audio/recording.ts
@@ -1,4 +1,5 @@
 import { Audio } from "expo-av";
+import * as FileSystem from "expo-file-system";
 
 export const audioInitalize = async () => {
   const response = await Audio.getPermissionsAsync();
@@ -6,4 +7,35 @@ export const audioInitalize = async () => {
   if (response.granted) return;
 
   await Audio.requestPermissionsAsync();
+};
+
+export async function startRecording() {
+  await Audio.setAudioModeAsync({
+    allowsRecordingIOS: true,
+    playsInSilentModeIOS: true,
+  });
+  const { recording } = await Audio.Recording.createAsync(
+    Audio.RecordingOptionsPresets.HIGH_QUALITY,
+    (recordingStatus) => {}
+  );
+  return recording;
+}
+
+export const postNewQuestion = async (recordFileURI: string) => {
+  const url = "https://hackday.kajilab.tk/upload";
+  const response = await FileSystem.uploadAsync(url, recordFileURI, {
+    uploadType: FileSystem.FileSystemUploadType.MULTIPART,
+    httpMethod: "POST",
+    headers: {
+      "Content-Type": "multipart/form-data",
+    },
+    fieldName: "file",
+    parameters: {
+      user_id: "6kSDkE0SNvWuVbiSVg8W",
+      is_answer: "false",
+      type: "questioner",
+    },
+  });
+  const newQuestion = JSON.parse(response.body);
+  return newQuestion;
 };
