+++ src/types/index.d.ts
@@ -20,7 +20,7 @@ export interface Question {
   user: User;
 }
 
-function isQuestion(obj: Object): obj is Question {
+export function isQuestion(obj: Object): obj is Question {
   let judge = true;
   [
     "question_id",
@@ -32,7 +32,7 @@ function isQuestion(obj: Object): obj is Question {
     "created_at",
     "user",
   ].forEach((key) => {
-    judge *= Object.prototype.hasOwnProperty(key);
+    judge = judge && Object.prototype.hasOwnProperty(key);
   });
   return judge;
 }
